# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "task index",
                "description": "get all tasks",
                "operationId": "tasksIndex",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Store Task",
                "description": "Store Task",
                "operationId": "storeTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "api_messages.store_task_successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/tasks/{task}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "show task",
                "description": "show task",
                "operationId": "showTask",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "show id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemNotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete task",
                "description": "Delete task request api",
                "operationId": "destroyTask",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessMessage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemNotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Task"
                ],
                "summary": "update task",
                "description": "update task request api",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessMessage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemNotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Task": {
                "title": "Task",
                "description": "Task model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "title of taask",
                        "type": "string",
                        "example": "string"
                    },
                    "decription": {
                        "title": "decription",
                        "description": "decription of taask",
                        "type": "string",
                        "example": "string"
                    },
                    "status": {
                        "title": "status",
                        "description": "status of task",
                        "type": "integer",
                        "example": "0"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Task"
                }
            },
            "StoreTaskRequest": {
                "title": "Store Task request",
                "description": "Store Task request body data",
                "required": [
                    "title",
                    "description",
                    "status"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "titl",
                        "type": "string",
                        "example": "new title"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string"
                    },
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateTaskRequest": {
                "title": "Update Task request",
                "description": "Update Task request body data",
                "required": [
                    "title",
                    "description",
                    "status"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "titl",
                        "type": "string",
                        "example": "new title"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string"
                    },
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "ItemNotFound": {
                "title": "Item Not Found",
                "description": "Item Not Found",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Item Not Found",
                        "type": "string",
                        "example": "Item Not Found"
                    }
                },
                "type": "object"
            },
            "Link": {
                "title": "Link",
                "description": "Link",
                "properties": {
                    "url": {
                        "title": "url",
                        "description": "url",
                        "type": "string",
                        "example": "http://127.0.0.1:8000/api/panel/users?page=1"
                    },
                    "label": {
                        "title": "label",
                        "description": "label",
                        "type": "string",
                        "example": "1"
                    },
                    "active": {
                        "title": "active",
                        "description": "active",
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
            "Meta": {
                "title": "Meta",
                "description": "Meta",
                "properties": {
                    "current_page": {
                        "title": "current_page",
                        "description": "current_page",
                        "type": "integer",
                        "example": "1"
                    },
                    "from": {
                        "title": "from",
                        "description": "from",
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "title": "last_page",
                        "description": "last_page",
                        "type": "integer",
                        "example": "5"
                    },
                    "links": {
                        "title": "links",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Link"
                        }
                    }
                },
                "type": "object"
            },
            "SuccessMessage": {
                "title": "Success",
                "description": "SuccessMessage",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "success message",
                        "type": "string",
                        "example": "success message"
                    }
                },
                "type": "object"
            },
            "TaskCollection": {
                "title": "Task Collection",
                "description": "Task list",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Task"
                        }
                    },
                    "links": {
                        "title": "links",
                        "description": "links",
                        "properties": {
                            "first": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/panel/users?page=2"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/panel/users?page=5"
                            },
                            "prev": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/panel/users?page=1"
                            },
                            "next": {
                                "type": "string",
                                "example": "http://127.0.0.1:8000/api/panel/users?page=3"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Meta"
                    },
                    "path": {
                        "title": "path",
                        "description": "path",
                        "type": "string",
                        "example": "http://127.0.0.1:8000/api/panel/users"
                    },
                    "per_page": {
                        "title": "per_page",
                        "description": "per_page",
                        "type": "integer",
                        "example": "10"
                    },
                    "to": {
                        "title": "to",
                        "description": "to",
                        "type": "integer",
                        "example": "10"
                    },
                    "total": {
                        "title": "total",
                        "description": "total",
                        "type": "integer",
                        "example": "20"
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "title": "Channel Resource",
                "description": "Get Channel",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Task"
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "title": "Unauthenticated",
                "description": "Unauthenticated",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "error message",
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}